requirements.txt
black==22.8.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cycler==0.11.0
flake8==5.0.4
fonttools==4.37.1
idna==3.3
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.2
packaging==21.3
pandas==1.4.4
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pyparsing==3.0.9
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
pytz==2022.2.1
pytorchtools==0.0.2
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
antlr4-python3-runtime==4.9.3 
hydra-core==1.2.0 
omegaconf==2.2.3
efficientnet-lite-pytorch==0.1.0
efficientnet-pytorch==0.7.0
librosa==0.9.2
seaborn==0.12.1